backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
chromadb>=0.4.15
sentence-transformers>=3.0.1
rank-bm25>=0.2.2
pymupdf>=1.24.0
anthropic>=0.7.7
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-readability>=0.1.3
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4
ollama>=0.3.0
# graphrag>=0.3.0  # Disabled: requires Python <3.13, incompatible with current environment
aiosqlite>=0.19.0
GitPython>=3.1.40